+++ src/contexts/ViewContext.js
@@ -1,5 +1,0 @@
-import { createContext } from 'react';
-
-const ViewContext = createContext();
-
-export default ViewContext;