+++ src/contexts/ViewsContext.js
@@ -1,0 +1,9 @@
+import { createContext } from 'react';
+
+const ViewsContext = createContext({
+  viewsState: {
+    views: [],
+  },
+});
+
+export default ViewsContext;