+++ src/hooks/__tests__/useViews.test.js
@@ -1,0 +1,14 @@
+import { renderHook } from '@testing-library/react-hooks';
+import useViews from '~/hooks/useViews';
+
+describe('[HOOK] useViews', () => {
+  it('should return views state', () => {
+    const { result } = renderHook(() => useViews());
+
+    expect(result.current).toEqual({
+      viewsState: {
+        views: [],
+      },
+    });
+  });
+});