+++ CHANGELOG.md
@@ -3,3 +3,9 @@
 ## 0.1.0
 
 - Initial version
+
+## 0.3.0
+
+- Add some tests
+- Create basic docs
+- Change `ViewContext` to `ViewsContext`